refac: move Friend layout to layouts folder
